mtd: nand: Add Hynix H27UCG8T2ETR-BC chip
This chip has a 6-byte ID according to spec.
A 5-byte ID was specified because Arasan NFC, which the chip is
added for, only reads 5 bytes of ID.
Additionally, the actual page size of the chip is 16K, but it
was changed to 8K because NFC doesn't support >8K page sizes.

Change-Id: Id70bb5f80597c3101359c3a7ae32413aac2e8c5c