fbdev: vpoutfb: Make check_var() conform to skeletonfb model
The API documentation for framebuffer devices doesn't explain
where to correct settings and where to return an error in check_var.
However, applications that call FBIOPUT_VSCREENINFO ioctl
only modify color OR timings in one increment
and expect a -EINVAL if it doesn't work.

Therefore, code was put in the driver to reject most incorrect
FBIOPUT_VSCREENINFO calls instead of silently correcting them
to an acceptable value.

As one intended effect, this makes checking if your modesetting
was successful extremely easy.

Change-Id: I749860487e2c34cfa6266e18afa31649134a5eef