fbdev: vpoutfb: Clarify video timings
Pixel clock speed was increased for 1024x768 in order to bring
every mode in the modedb to similar FPS, which are all around 56-60.

Mode list was updated to better visualize what modes are available.
The FPS in the modename represents the original intent by CEA or VESA
standard bodies, while the FPS in the comments represents the real
FPS as calculated from the timings.

Change-Id: I94c9695956388555b0ecac0fb973cb1847e8938e