nand: arasan_nfc: Calculate ecc_addr based on flash params
This commit makes ECC always align to the end of OOB, allowing
filesystems such as JFFS2 to use OOB in a fuller fashion,
as well as fixing a discrepancy in the reported OOB alignment
and the actual way OOB is structured.

However, it also means that pages written before this commit
will fail to read until rewritten.

Change-Id: Id3e1fb1b50fde679360db566f212df2de428e125