ARM: mcom: Rewrite code for CPU1 booting
Currently CPU1 is waiting in a loop in U-Boot SPL to be released by
Linux. This is an incorrect approach since the OS may not use the
second core.

The proper way to deal with CPU1 is to turn it off early in U-Boot
and turn it on in Linux.

This commit boots CPU1 by enabling its power domain. Prior to that
BOOT_REMAP register is set to SPRAM where the startup code is loaded.
CPU1 changes BOOT_REMAP back to NORMPORT early in Linux.

Issue: #MCOM02SW-122
Change-Id: Iadba29d7de9980cce10123a0cf5733cabe1bba2e