drm/vpout: Add labels lookup mechanism for connectors
DRM core generated names for the connectors using an index
that depended on initialization order. In some cases an order
may change, it can lead to undesired behavior if we want
to use this name in kernel command line options.

This change provides abbility to append a label from DTS
for name and it will look like:

  LVDS-<order_idx>-<label>

The old part LVDS-<order_idx> is needed because it's yet
showing in some applications.

Now the kernel command-line option will look like:

  video=<label>:<xres>x<yres>

This format is needed because connector label is constant.
If a label isn't used then old format available, but should
 be used with caution.

Issue: #MCOM02SW-949
Change-Id: I4456fde41265a08b6f9abdc3a7729d6a3fccea16